Pull request for issue #8

+++ routes/players.js
@@ -1,5 +1,6 @@
 const moment = require('moment');
 const express = require('express');
+const mongoose = require('mongoose'); // Added for ObjectId validation
 const { Team } = require('../models/team');
 const { Player, validatePlayer } = require('../models/player');
 
@@ -24,9 +25,8 @@ router.post('/', async (req, res) => {
       _id: team._id,
       name: team.name,
     },
-    playerCount: req.body.playerCount,
+    loanDaysRemaining: req.body.loanDaysRemaining,
     loanCost: req.body.loanCost,
-    timeStamp: moment.toJSON(),
   });
 
   await player.save();
@@ -34,10 +34,14 @@ router.post('/', async (req, res) => {
 });
 
 router.put('/:id', async (req, res) => {
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    res.status(400).send('Invalid Player ID.');
+    return;
+  }
   const { error } = validatePlayer(req.body);
   if (error) return res.status(400).send(error.details[0].message);
 
-  const team = Team.findById(req.body.teamId);
+  const team = await Team.findById(req.body.teamId);
   if (!team) return res.status(400).send('Invalid Team details provided');
 
   const player = await Player.findByIdAndUpdate(
@@ -48,7 +52,7 @@ router.put('/:id', async (req, res) => {
         _id: team._id,
         name: team.name,
       },
-      playerCount: req.body.playerCount,
+      loanDaysRemaining: req.body.loanDaysRemaining,
       loanCost: req.body.loanCost,
     },
     {
@@ -61,14 +65,22 @@ router.put('/:id', async (req, res) => {
 });
 
 router.delete('/:id', async (req, res) => {
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    res.status(400).send('Invalid Player ID.');
+    return;
+  }
   const player = await Player.findByIdAndRemove(req.params.id);
   if (!player) return res.status(404).send('Player ID not found');
   res.send(player);
 });
 
 router.get('/:id', async (req, res) => {
+  if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+    res.status(400).send('Invalid Player ID.');
+    return;
+  }
   const player = await Player.findById(req.params.id);
-  if (!player) return res.status(404).send('await Player ID not found');
+  if (!player) return res.status(404).send('Player with given ID was not found.');
   res.send(player);
 });
